+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [0.0.10] / 2021-12-22
+- Add [Parameter] again... (null error)
+- Fix Vendor `null` error
 - Remove [Parameter]
 - Add VendorDescription
 - Add VendorId